refactor: fastify types
